Restructure browser tests
