+++ browserTests/generic/titleBarTest.js
