+++ browserTests/views/addressTest.js
@@ -2,7 +2,7 @@
 import { Selector } from 'testcafe';
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 
-import config from './config';
+import config from '../config';
 const { server } = config;
 fixture`Address view test`
   .page`http://${server.address}:${server.port}/fi/address/turku/Linnankatu/14`
