+++ browserTests/views/areaTest.js
@@ -2,7 +2,7 @@
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 import { ClientFunction, Selector } from 'testcafe';
 
-import config from './config';
+import config from '../config';
 const { server } = config;
 
 fixture`Area view test`
