+++ browserTests/views/homeTest.js
@@ -2,8 +2,8 @@
 import { Selector, ClientFunction } from 'testcafe';
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 
-import config from './config';
-import PaperButtonTest from '../src/components/PaperButton/__testCafe__/PaperButton.tc';
+import config from '../config';
+import PaperButtonTest from '../../src/components/PaperButton/__testCafe__/PaperButton.tc';
 const { server } = config;
 
 fixture`Home view test`
