+++ browserTests/views/mobilityPlatformTest.js
@@ -2,7 +2,7 @@
 import { Selector, ClientFunction } from 'testcafe';
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 
-import config from './config';
+import config from '../config';
 const { server } = config;
 
 fixture`Mobility map view test`.page`http://${server.address}:${server.port}/fi/mobility`.beforeEach(async () => {
