+++ browserTests/views/serviceTest.js
@@ -1,6 +1,6 @@
 import { waitForReact, ReactSelector } from 'testcafe-react-selectors';
 import { ClientFunction, Selector } from 'testcafe';
-import config from './config';
+import config from '../config';
 
 const { server } = config;
